Switch to NAN.  Welcome to 0.5.0.
Take the opportunity to clean up the API in a couple of places.  i2cRead
and i2cWrite now take a buffer argument, and return the exit status of
the underlying bcm2835 functions to indicate possible failure.  Also add
an spiWrite function for write-only SPI transfers.

Switching to NAN now means node-rpio works on all release from 0.8 to
the latest 5.2.0.