Add support for the Compute Module 3.
Submitted by Felix Galindo in #49.  I'm still not entirely sure this is
correct as with expansion boards the pinouts will be different as well
as having many more pins available, but other users in #48 have reported
this works, so let's just enable it for now and clean up later if
necessary.

Release 1.0.1.