Always use device tree, remove cpuinfo detection.
This is required for Raspberry Pi 2 and newer anyway so the vast
majority of users will already have this, and it's easy to enable for
anyone still using really old hardware and OS releases.

Not having to manually keep track of every hardware id should enable
support across all boards, and reduce friction for adding new ones.