Support switching to input mode in read{,buf}()
This improves support for timing-critical devices such as the DHT11,
where we want to read from the device as soon after switching to input
mode as possible.  Enabling input mode in the same call as the read
avoids the lengthy JavaScript function call overhead of around 100us
which can be the difference between a successful read or not.