Overhaul and improve the DHT11 example.
This uses the new mode option for readbuf() which greatly reduces the
delay between switching to input mode and when we start reading from the
device.  Also fixes a bunch of incorrect assumptions about the protocol
after re-reading the datasheet, and improves the quality of the results
based on some real-world usage.

Seems to work pretty well in testing on Raspberry Pi 1 and 3 models.